added custom hook forjob title
