+++ client/src/components/Banner/Banner.js
@@ -1,58 +1,16 @@
-import {useEffect, useState} from 'react';
 import {Container,Row,Col} from 'react-bootstrap';
 import {ArrowRightCircle} from 'react-bootstrap-icons';
 import 'animate.css';
 import headerImg from "../../assets/img/header-img.svg";
 import TrackVisibility from 'react-on-screen';
 import {aboutme} from '../../common/constants';
+import { useJobTitle } from '../../hooks/useJobTitle';
 import './Banner.scss';
 
+
 export const Banner = ({contactRef})=>{
-    const period = 350;
-    const deletingPeriod = 100;
-    const [loopNum,setLoopNum] = useState(0);
-    const [isDeleting, setIsDeleting] = useState(false);
-    const [delta,setDelta] = useState(period);
-    const [text,setText] = useState('');
-    const toRotate = ['Web Developer', "Software Developer", "Full-Stack Developer"];
-    // const [index, setIndex] = useState(1);
-
-  // console.log('redering')
-    useEffect(()=>{
-        let ticker = setInterval(()=>{
-            tick()
-        },delta)
-
-        return ()=>{ clearInterval(ticker)};
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    },[text]);
-
-    const tick = ()=>{
-      
-        // console.log('running',text)
-        let i = loopNum % toRotate.length;
-        let fullText = toRotate[i];
-        let updatedText = isDeleting ? fullText.substring(0,text.length-1) : fullText.substring(0,text.length + 1);
-
-        setText(updatedText);
-
-        if(isDeleting){
-            setDelta(deletingPeriod);
-        }
-
-        if(!isDeleting && updatedText === fullText){
-            setIsDeleting(true);
-            setDelta(period);
-        } else if(isDeleting && updatedText === ''){
-            setIsDeleting(false);
-            setLoopNum(loopNum + 1);
-            // setIndex(1);
-            setDelta(period);
-          } else {
-            // setIndex(prevIndex => prevIndex + 1);
-          }
-    }
-
+    
+  const text = useJobTitle();
     return (
         <section className="banner" id="home">
         <Container>
